Support commit messages in Git submodules
In recent Git versions, the .git directories of submodules are not
stored in the submodule itself, but rather in a subdirectory of the .git
directory of the parent directory, e.g. the submodule foo/bar has its
.git directory in .git/modules/foo/bar/ of the parent repository.

Change the regular expression for editable Git files to allow for
arbitrary nesting, such that commit messages in submodules are
considered editable as well.